style: lint, format and restructure files
