+++ wallet/.flake8
@@ -1,0 +1,5 @@
+[flake8]
+max-line-length= 127
+ignore = E203,W605,W503
+max-complexity= 16
+exclude = .git,__pycache__,__init__.py,.mypy_cache,.pytest_cache
