+++ wallet/PassInformation/__init__.py
@@ -1,0 +1,1 @@
+from .PassInformation import PassInformation
