+++ wallet/PassProps/Alignment.py
@@ -1,0 +1,8 @@
+class Alignment:
+    """Text Alignment"""
+
+    LEFT = "PKTextAlignmentLeft"
+    CENTER = "PKTextAlignmentCenter"
+    RIGHT = "PKTextAlignmentRight"
+    JUSTIFIED = "PKTextAlignmentJustified"
+    NATURAL = "PKTextAlignmentNatural"
