+++ wallet/PassProps/DateStyle.py
@@ -1,0 +1,8 @@
+class DateStyle:
+    """Date Style"""
+
+    NONE = "PKDateStyleNone"
+    SHORT = "PKDateStyleShort"
+    MEDIUM = "PKDateStyleMedium"
+    LONG = "PKDateStyleLong"
+    FULL = "PKDateStyleFull"
