+++ wallet/PassProps/NumberStyle.py
@@ -1,0 +1,7 @@
+class NumberStyle:
+    """Number Style"""
+
+    DECIMAL = "PKNumberStyleDecimal"
+    PERCENT = "PKNumberStylePercent"
+    SCIENTIFIC = "PKNumberStyleScientific"
+    SPELLOUT = "PKNumberStyleSpellOut"
