+++ wallet/PassProps/TransitType.py
@@ -1,0 +1,8 @@
+class TransitType:
+    """Transit Type for Boarding Passes"""
+
+    AIR = "PKTransitTypeAir"
+    TRAIN = "PKTransitTypeTrain"
+    BUS = "PKTransitTypeBus"
+    BOAT = "PKTransitTypeBoat"
+    GENERIC = "PKTransitTypeGeneric"
