+++ wallet/PassProps/__init__.py
@@ -1,0 +1,8 @@
+from .Alignment import Alignment
+from .Barcode import Barcode
+from .DateStyle import DateStyle
+from .Field import Field
+from .IBeacon import IBeacon
+from .Location import Location
+from .NumberStyle import NumberStyle
+from .TransitType import TransitType
