+++ wallet/PassStyles/Coupon.py
@@ -1,0 +1,9 @@
+from wallet.PassInformation import PassInformation
+
+
+class Coupon(PassInformation):
+    """Wallet Coupon Pass"""
+
+    def __init__(self):
+        super(Coupon, self).__init__()
+        self.jsonname = "coupon"
