+++ wallet/PassStyles/__init__.py
@@ -1,0 +1,5 @@
+from .BoardingPass import BoardingPass
+from .Coupon import Coupon
+from .EventTicket import EventTicket
+from .Generic import Generic
+from .StoreCard import StoreCard
