+++ wallet/exceptions.py
@@ -2,6 +2,7 @@
 Wallet Exceptions
 """
 
+
 class PassParameterException(Exception):
     """
     Parameter based Exception
